// src/app/shared/models/Shipping.ts
export default class Shipping {
    id?: number;
    firstName: string;
    lastName: string;
    city: string;
    area: string;
    street: string;
    buildingNumber: string;
    floorNumber: string;
    appartmentNumber: string;
    landmark: string;
    locationType: string;
    mobile: string;
    landLine: string;
    note: string;
    countryId: number;
}